pyproject.toml
[tool.poetry]
name = "cow-dataviewer"
version = "0.1.0"
description = ""
authors = ["Mamoru Matsunaga <mamoru74521103@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^0.75.0"
opencv-python = "^4.5.1"
PyYAML = "^5.4.1"
matplotlib = "^3.3.4"
black = "^20.8b1"
flake8 = "^3.8.4"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
